Move alter blob table tests to AlterTablePlanTest
Some of the tests used a low level `getTableParameter` directly instead
of going through the higher level methods - making refactorings more
painful as they break all these tests.
