+++ server/src/test/java/io/crate/planner/node/ddl/AlterTablePlanTest.java
@@ -24,16 +24,18 @@ package io.crate.planner.node.ddl;
 import static io.crate.replication.logical.LogicalReplicationSettings.REPLICATION_SUBSCRIPTION_NAME;
 import static io.crate.testing.TestingHelpers.createNodeContext;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.IOException;
 
-import org.assertj.core.api.Assertions;
+import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.common.settings.Settings;
 import org.junit.Before;
 import org.junit.Test;
 
 import io.crate.analyze.BoundAlterTable;
 import io.crate.data.Row;
+import io.crate.exceptions.OperationOnInaccessibleRelationException;
 import io.crate.metadata.CoordinatorTxnCtx;
 import io.crate.planner.operators.SubQueryResults;
 import io.crate.test.integration.CrateDummyClusterServiceUnitTest;
@@ -68,10 +70,75 @@ public class AlterTablePlanTest extends CrateDummyClusterServiceUnitTest {
 
     @Test
     public void test_alter_forbidden_settings_on_a_replicated_table() throws IOException {
-        Assertions.assertThatThrownBy(() -> analyze("Alter table doc.test set(number_of_shards = 1)"))
+        assertThatThrownBy(() -> analyze("Alter table doc.test set(number_of_shards = 1)"))
             .isExactlyInstanceOf(IllegalArgumentException.class)
             .hasMessageContaining("Invalid property \"number_of_shards\" passed to [ALTER | CREATE] TABLE statement");
     }
+
+    @Test
+    public void testAlterBlobTableWithInvalidProperty() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        assertThatThrownBy(() -> analyze("alter blob table blobs set (foobar='2')"))
+            .isExactlyInstanceOf(IllegalArgumentException.class)
+            .hasMessage("Invalid property \"foobar\" passed to [ALTER | CREATE] TABLE statement");
+    }
+
+    @Test
+    public void testAlterBlobTableWithReplicas() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        BoundAlterTable alterTable = analyze("alter blob table blobs set (number_of_replicas=2)");
+        assertThat(alterTable.table().ident().name()).isEqualTo("blobs");
+        assertThat(alterTable.settings().getAsInt(IndexMetadata.SETTING_NUMBER_OF_REPLICAS, 0))
+            .isEqualTo(2);
+    }
+
+    @Test
+    public void test_alter_setting_block_read_only() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        BoundAlterTable alterTable = analyze("alter blob table blobs set (\"blocks.read_only_allow_delete\"=true)");
+        assertThat(alterTable.table().ident().name()).isEqualTo("blobs");
+        assertThat(alterTable.settings().getAsBoolean(IndexMetadata.SETTING_READ_ONLY_ALLOW_DELETE, false)).isTrue();
+    }
+
+    @Test
+    public void testAlterBlobTableWithPath() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        assertThatThrownBy(() -> analyze("alter blob table blobs set (blobs_path=1)"))
+            .isExactlyInstanceOf(IllegalArgumentException.class)
+            .hasMessage("Invalid property \"blobs_path\" passed to [ALTER | CREATE] TABLE statement");
+    }
+
+    @Test
+    public void testAlterBlobTableRenameTable() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        assertThatThrownBy(() -> e.analyze("alter blob table blobs rename to blobbier"))
+            .isExactlyInstanceOf(OperationOnInaccessibleRelationException.class)
+            .hasMessage("The relation \"blob.blobs\" doesn't support or allow ALTER RENAME operations");
+    }
+
+    @Test
+    public void testAlterBlobTableRenameTableWithExplicitSchema() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        assertThatThrownBy(() -> e.analyze("alter blob table schema.blobs rename to blobbier"))
+            .isExactlyInstanceOf(IllegalArgumentException.class)
+            .hasMessage("The Schema \"schema\" isn't valid in a [CREATE | ALTER] BLOB TABLE clause");
+    }
+
+    @Test
+    public void testAlterBlobTableOpenClose() throws Throwable {
+        e.addBlobTable("create blob table blobs");
+        assertThatThrownBy(() -> e.analyze("alter blob table blobs close"))
+            .isExactlyInstanceOf(OperationOnInaccessibleRelationException.class)
+            .hasMessage("The relation \"blob.blobs\" doesn't support or allow ALTER CLOSE operations");
+    }
+
+    @Test
+    public void testAlterBlobTableOpenCloseWithExplicitSchema() {
+        assertThatThrownBy(() -> e.analyze("alter blob table schema.blob close"))
+            .isExactlyInstanceOf(IllegalArgumentException.class)
+            .hasMessage("The Schema \"schema\" isn't valid in a [CREATE | ALTER] BLOB TABLE clause");
+    }
+
 
     private BoundAlterTable analyze(String stmt) {
         AlterTablePlan plan = e.plan(stmt);
